app/requirements.txt
streamlit
openai-whisper
ffmpeg-python
torch
numpy
pandas
scikit-learn